+++ custom_components/hovalconnect/const.py
@@ -6,6 +6,9 @@ API_BASE  = "https://azure-iot-prod.hoval.com/core"
 AUTH_BASE = "https://akwc5scsc.accounts.ondemand.com"
 AUTH_TOKEN_URL = f"{AUTH_BASE}/oauth2/token"
 CLIENT_ID = "991b54b2-7e67-47ef-81fe-572e21c59899"
+DEFAULT_FRONTEND_APP_VERSION = "3.2.0"
+FRONTEND_VERSION_CHECK_INTERVAL_SECONDS = 6 * 60 * 60
+GOOGLE_PLAY_APP_URL = "https://play.google.com/store/apps/details?id=com.hoval.connect2&hl=en_US&gl=US"
 
 API_MY_PLANTS      = f"{API_BASE}/api/my-plants?page=0&size=12"
 API_PLANT_SETTINGS = f"{API_BASE}/v1/plants/{{plant_id}}/settings"